[IMP] service unit test

+++ src/test/groovy/io/choerodon/devops/api/controller/v1/DevopsServiceControllerSpec.groovy
@@ -4,31 +4,35 @@ import io.choerodon.core.domain.Page
 import io.choerodon.devops.IntegrationTestConfiguration
 import io.choerodon.devops.api.dto.DevopsServiceDTO
 import io.choerodon.devops.api.dto.DevopsServiceReqDTO
+import io.choerodon.devops.api.dto.iam.ProjectWithRoleDTO
+import io.choerodon.devops.api.dto.iam.RoleDTO
 import io.choerodon.devops.domain.application.entity.PortMapE
-import io.choerodon.devops.domain.application.entity.ProjectE
-import io.choerodon.devops.domain.application.entity.UserAttrE
-import io.choerodon.devops.domain.application.entity.gitlab.GitlabGroupE
-import io.choerodon.devops.domain.application.entity.gitlab.GitlabMemberE
 import io.choerodon.devops.domain.application.repository.*
-import io.choerodon.devops.domain.application.valueobject.Organization
+import io.choerodon.devops.domain.application.valueobject.RepositoryFile
 import io.choerodon.devops.infra.common.util.EnvUtil
 import io.choerodon.devops.infra.common.util.FileUtil
 import io.choerodon.devops.infra.common.util.GitUtil
 import io.choerodon.devops.infra.common.util.enums.AccessLevel
 import io.choerodon.devops.infra.dataobject.*
+import io.choerodon.devops.infra.dataobject.gitlab.MemberDO
+import io.choerodon.devops.infra.dataobject.iam.OrganizationDO
+import io.choerodon.devops.infra.dataobject.iam.ProjectDO
+import io.choerodon.devops.infra.feign.GitlabServiceClient
+import io.choerodon.devops.infra.feign.IamServiceClient
 import io.choerodon.devops.infra.mapper.*
 import io.choerodon.websocket.helper.EnvListener
+import org.mockito.Mockito
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.web.client.TestRestTemplate
 import org.springframework.context.annotation.Import
-import org.springframework.http.HttpEntity
-import org.springframework.http.HttpHeaders
-import org.springframework.http.MediaType
+import org.springframework.http.*
+import spock.lang.Shared
 import spock.lang.Specification
 import spock.lang.Stepwise
 
+import static org.mockito.Matchers.*
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT
 
 /**
@@ -43,7 +47,6 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
 @Stepwise
 class DevopsServiceControllerSpec extends Specification {
 
-    private static flag = 0
     private static id = 0
 
     @Autowired
@@ -80,73 +83,119 @@ class DevopsServiceControllerSpec extends Specification {
     @Autowired
     @Qualifier("mockEnvListener")
     private EnvListener envListener
+
+    @Shared
+    DevopsEnvCommandDO devopsEnvCommandDO = new DevopsEnvCommandDO()
+    @Shared
+    DevopsEnvFileResourceDO devopsEnvFileResourceDO = new DevopsEnvFileResourceDO()
+    @Shared
+    DevopsServiceAppInstanceDO devopsServiceAppInstanceDO = new DevopsServiceAppInstanceDO()
+    @Shared
+    ApplicationInstanceDO applicationInstanceDO = new ApplicationInstanceDO()
+    @Shared
+    DevopsEnvironmentDO devopsEnvironmentDO = new DevopsEnvironmentDO()
+
     @Autowired
-    @Qualifier("mockIamRepository")
     private IamRepository iamRepository
     @Autowired
-    @Qualifier("mockGitlabRepository")
     private GitlabRepository gitlabRepository
     @Autowired
-    @Qualifier("mockUserAttrRepository")
-    private UserAttrRepository userAttrRepository
+    private GitlabProjectRepository gitlabProjectRepository
     @Autowired
-    @Qualifier("mockGitlabGroupMemberRepository")
     private GitlabGroupMemberRepository gitlabGroupMemberRepository
 
-    private DevopsEnvCommandDO devopsEnvCommandDO
-    private DevopsEnvFileResourceDO devopsEnvFileResourceDO
-    private DevopsServiceAppInstanceDO devopsServiceAppInstanceDO
+    IamServiceClient iamServiceClient = Mockito.mock(IamServiceClient.class)
+    GitlabServiceClient gitlabServiceClient = Mockito.mock(GitlabServiceClient.class)
+
+    def setupSpec() {
+        FileUtil.copyFile("src/test/gitops/org/pro/env/test-svc.yaml", "gitops/org/pro/env")
+
+        devopsEnvCommandDO.setId(1L)
+
+        devopsEnvironmentDO.setId(1L)
+        devopsEnvironmentDO.setProjectId(1L)
+        devopsEnvironmentDO.setActive(true)
+        devopsEnvironmentDO.setGitlabEnvProjectId(1L)
+        devopsEnvironmentDO.setSequence(1L)
+        devopsEnvironmentDO.setCode("env")
+        devopsEnvironmentDO.setDevopsEnvGroupId(1L)
+
+        applicationInstanceDO.setId(1L)
+        applicationInstanceDO.setProjectId(1L)
+        applicationInstanceDO.setCode("test")
+        applicationInstanceDO.setAppId(1L)
+        applicationInstanceDO.setEnvId(1L)
+        applicationInstanceDO.setAppVersionId(1L)
+
+        devopsEnvFileResourceDO.setId(1L)
+        devopsEnvFileResourceDO.setEnvId(1L)
+        devopsEnvFileResourceDO.setResourceId(1L)
+        devopsEnvFileResourceDO.setResourceType("Service")
+        devopsEnvFileResourceDO.setFilePath("test-svc.yaml")
+
+        devopsServiceAppInstanceDO.setId(1L)
+        devopsServiceAppInstanceDO.setServiceId(1L)
+        devopsServiceAppInstanceDO.setAppInstanceId(1L)
+    }
 
     def setup() {
-        if (flag == 0) {
-            FileUtil.copyFile("src/test/gitops/org/pro/env/test-svc.yaml", "gitops/org/pro/env")
+        iamRepository.initMockIamService(iamServiceClient)
+        gitlabRepository.initMockService(gitlabServiceClient)
+        gitlabProjectRepository.initMockService(gitlabServiceClient)
+        gitlabGroupMemberRepository.initMockService(gitlabServiceClient)
 
-            devopsEnvCommandDO = new DevopsEnvCommandDO()
-            devopsEnvCommandDO.setId(1L)
-            devopsEnvCommandMapper.insert(devopsEnvCommandDO)
+        ProjectDO projectDO = new ProjectDO()
+        projectDO.setId(1L)
+        projectDO.setCode("pro")
+        projectDO.setOrganizationId(1L)
+        ResponseEntity<ProjectDO> responseEntity = new ResponseEntity<>(projectDO, HttpStatus.OK)
+        Mockito.doReturn(responseEntity).when(iamServiceClient).queryIamProject(1L)
 
-            DevopsEnvironmentDO devopsEnvironmentDO = new DevopsEnvironmentDO()
-            devopsEnvironmentDO.setId(1L)
-            devopsEnvironmentDO.setProjectId(1L)
-            devopsEnvironmentDO.setActive(true)
-            devopsEnvironmentDO.setGitlabEnvProjectId(1L)
-            devopsEnvironmentDO.setSequence(1L)
-            devopsEnvironmentDO.setCode("env")
-            devopsEnvironmentDO.setDevopsEnvGroupId(1L)
-            devopsEnvironmentMapper.insert(devopsEnvironmentDO)
+        OrganizationDO organizationDO = new OrganizationDO()
+        organizationDO.setId(1L)
+        organizationDO.setCode("org")
+        ResponseEntity<OrganizationDO> responseEntity1 = new ResponseEntity<>(organizationDO, HttpStatus.OK)
+        Mockito.doReturn(responseEntity1).when(iamServiceClient).queryOrganizationById(1L)
 
-            ApplicationInstanceDO applicationInstanceDO = new ApplicationInstanceDO()
-            applicationInstanceDO.setId(1L)
-            applicationInstanceDO.setProjectId(1L)
-            applicationInstanceDO.setCode("test")
-            applicationInstanceDO.setAppId(1L)
-            applicationInstanceDO.setEnvId(1L)
-            applicationInstanceDO.setAppVersionId(1L)
-            applicationInstanceMapper.insert(applicationInstanceDO)
+        MemberDO memberDO = new MemberDO()
+        memberDO.setAccessLevel(AccessLevel.OWNER)
+        ResponseEntity<MemberDO> responseEntity2 = new ResponseEntity<>(memberDO, HttpStatus.OK)
+        Mockito.when(gitlabServiceClient.getUserMemberByUserId(anyInt(), anyInt())).thenReturn(responseEntity2)
 
-            devopsEnvFileResourceDO = new DevopsEnvFileResourceDO()
-            devopsEnvFileResourceDO.setId(1L)
-            devopsEnvFileResourceDO.setEnvId(1L)
-            devopsEnvFileResourceDO.setResourceId(1L)
-            devopsEnvFileResourceDO.setResourceType("Service")
-            devopsEnvFileResourceDO.setFilePath("test-svc.yaml")
-            devopsEnvFileResourceMapper.insert(devopsEnvFileResourceDO)
+        List<RoleDTO> roleDTOList = new ArrayList<>()
+        RoleDTO roleDTO = new RoleDTO()
+        roleDTO.setCode("role/project/default/project-owner")
+        roleDTOList.add(roleDTO)
+        List<ProjectWithRoleDTO> projectWithRoleDTOList = new ArrayList<>()
+        ProjectWithRoleDTO projectWithRoleDTO = new ProjectWithRoleDTO()
+        projectWithRoleDTO.setName("pro")
+        projectWithRoleDTO.setRoles(roleDTOList)
+        projectWithRoleDTOList.add(projectWithRoleDTO)
+        Page<ProjectWithRoleDTO> projectWithRoleDTOPage = new Page<>()
+        projectWithRoleDTOPage.setContent(projectWithRoleDTOList)
+        projectWithRoleDTOPage.setTotalPages(2)
+        ResponseEntity<Page<ProjectWithRoleDTO>> pageResponseEntity = new ResponseEntity<>(projectWithRoleDTOPage, HttpStatus.OK)
+        Mockito.doReturn(pageResponseEntity).when(iamServiceClient).listProjectWithRole(anyLong(), anyInt(), anyInt())
 
-            devopsServiceAppInstanceDO = new DevopsServiceAppInstanceDO()
-            devopsServiceAppInstanceDO.setId(1L)
-            devopsServiceAppInstanceDO.setServiceId(1L)
-            devopsServiceAppInstanceDO.setAppInstanceId(1L)
-            devopsServiceAppInstanceMapper.insert(devopsServiceAppInstanceDO)
-
-            flag = 1
-        }
+        RepositoryFile repositoryFile = new RepositoryFile()
+        repositoryFile.setFilePath("testFilePath")
+        ResponseEntity<RepositoryFile> responseEntity3 = new ResponseEntity<>(repositoryFile, HttpStatus.OK)
+        Mockito.when(gitlabServiceClient.createFile(anyInt(), anyString(), anyString(), anyString(), anyInt())).thenReturn(responseEntity3)
+        Mockito.when(gitlabServiceClient.updateFile(anyInt(), anyString(), anyString(), anyString(), anyInt())).thenReturn(responseEntity3)
     }
 
     def "CheckName"() {
-        when:
+        given: '插入数据'
+        devopsEnvCommandMapper.insert(devopsEnvCommandDO)
+        devopsEnvironmentMapper.insert(devopsEnvironmentDO)
+        applicationInstanceMapper.insert(applicationInstanceDO)
+        devopsEnvFileResourceMapper.insert(devopsEnvFileResourceDO)
+        devopsServiceAppInstanceMapper.insert(devopsServiceAppInstanceDO)
+
+        when: '检查网络唯一性'
         def exist = restTemplate.getForObject("/v1/projects/1/service/check?envId=1&name=svc", Boolean.class)
 
-        then:
+        then: '校验返回值'
         exist
     }
 
@@ -168,39 +217,17 @@ class DevopsServiceControllerSpec extends Specification {
         devopsServiceReqDTO.setPorts(portMapES)
         devopsServiceReqDTO.setExternalIp("1.1.1.1")
 
-        UserAttrE userAttrE = new UserAttrE(1L, 1L)
-
-        GitlabGroupE gitlabGroupE = new GitlabGroupE()
-        gitlabGroupE.setDevopsEnvGroupId(1)
-
-        GitlabMemberE gitlabMemberE = new GitlabMemberE()
-        gitlabMemberE.setAccessLevel(AccessLevel.OWNER.toValue())
-
-        Organization organization = new Organization()
-        organization.setId(1L)
-        organization.setCode("org")
-
-        ProjectE projectE = new ProjectE()
-        projectE.setId(1L)
-        projectE.setCode("pro")
-        projectE.setOrganization(organization)
-
         envUtil.checkEnvConnection(_ as Long, _ as EnvListener) >> null
-        userAttrRepository.queryById(_ as Long) >> userAttrE
-        devopsProjectRepository.queryDevopsProject(_ as Long) >> gitlabGroupE
-        gitlabGroupMemberRepository.getUserMemberByUserId(_ as Integer, _ as Integer) >> gitlabMemberE
-        iamRepository.queryIamProject(_ as Long) >> projectE
-        iamRepository.queryOrganizationById(_ as Long) >> organization
 
-        when:
+        when: '部署网络'
         def result = restTemplate.postForObject("/v1/projects/1/service", devopsServiceReqDTO, Boolean.class)
 
-        then:
+        then: '校验返回值'
         result
     }
 
     def "Update"() {
-        given:
+        given: 'mock envUtil'
         List<PortMapE> portMapES = new ArrayList<>()
         PortMapE portMapE = new PortMapE()
         portMapE.setPort(7777L)
@@ -219,125 +246,63 @@ class DevopsServiceControllerSpec extends Specification {
         newDevopsServiceReqDTO.setPorts(portMapES)
         newDevopsServiceReqDTO.setExternalIp("1.2.1.1")
 
-        UserAttrE userAttrE = new UserAttrE(1, 1)
-
-        GitlabGroupE gitlabGroupE = new GitlabGroupE()
-        gitlabGroupE.setDevopsEnvGroupId(1)
-
-        GitlabMemberE gitlabMemberE
-        gitlabMemberE = new GitlabMemberE()
-        gitlabMemberE.setAccessLevel(AccessLevel.OWNER.toValue())
-
-        Organization organization = new Organization()
-        organization.setId(1L)
-        organization.setCode("org")
-
-        ProjectE projectE = new ProjectE()
-        projectE.setId(1L)
-        projectE.setCode("pro")
-        projectE.setOrganization(organization)
         envUtil.checkEnvConnection(_ as Long, _ as EnvListener) >> null
-        userAttrRepository.queryById(_ as Long) >> userAttrE
-        devopsProjectRepository.queryDevopsProject(_ as Long) >> gitlabGroupE
-        gitlabGroupMemberRepository.getUserMemberByUserId(_ as Integer, _ as Integer) >> gitlabMemberE
-        iamRepository.queryIamProject(_ as Long) >> projectE
-        iamRepository.queryOrganizationById(_ as Long) >> organization
         id = devopsServiceRepository.selectByNameAndEnvId("svcsvc", 1L).getId()
         devopsEnvFileResourceDO = devopsEnvFileResourceMapper.selectByPrimaryKey(1L)
         devopsEnvFileResourceDO.setResourceId(id)
         devopsEnvFileResourceMapper.updateByPrimaryKey(devopsEnvFileResourceDO)
 
-        when:
+        when: '更新网络'
         restTemplate.put("/v1/projects/1/service/{id}", newDevopsServiceReqDTO, id)
 
-        then:
-        devopsServiceMapper.selectByPrimaryKey(id).getExternalIp().equals("1.2.1.1")
+        then: '校验返回值'
+        devopsServiceMapper.selectByPrimaryKey(id).getExternalIp() == "1.2.1.1"
     }
 
     def "Delete"() {
-        given:
-        Organization organization = new Organization()
-        organization.setId(1L)
-        organization.setCode("org")
-
-        ProjectE projectE = new ProjectE()
-        projectE.setId(1L)
-        projectE.setCode("pro")
-        projectE.setOrganization(organization)
-
-        UserAttrE userAttrE = new UserAttrE(1, 1)
-
-        GitlabGroupE gitlabGroupE = new GitlabGroupE()
-        gitlabGroupE.setDevopsEnvGroupId(1)
-
-        GitlabMemberE gitlabMemberE
-        gitlabMemberE = new GitlabMemberE()
-        gitlabMemberE.setAccessLevel(AccessLevel.OWNER.toValue())
+        given: 'mock envUtil'
         envUtil.checkEnvConnection(_ as Long, _ as EnvListener) >> null
-        iamRepository.queryIamProject(_ as Long) >> projectE
-        iamRepository.queryOrganizationById(_ as Long) >> organization
-        userAttrRepository.queryById(_ as Long) >> userAttrE
-        devopsProjectRepository.queryDevopsProject(_ as Long) >> gitlabGroupE
-        gitlabGroupMemberRepository.getUserMemberByUserId(_ as Integer, _ as Integer) >> gitlabMemberE
-        gitlabRepository.deleteFile(_ as Integer, _ as String, _ as String, _ as Integer) >> null
 
-        when:
+        ResponseEntity responseEntity = new ResponseEntity<>(HttpStatus.OK)
+        Mockito.when(gitlabServiceClient.deleteFile(anyInt(), anyString(), anyString(), anyInt())).thenReturn(responseEntity)
+
+        when: '删除网络'
         restTemplate.delete("/v1/projects/1/service/{id}", id)
 
-        then:
-        devopsEnvCommandRepository.queryByObject("service", id).getCommandType().equals("delete")
-    }
-
-    def "PageByOptions"() {
-        given:
-        String infra = "{\"searchParam\":{\"name\":[\"svcsvc\"]}}"
-
-        HttpHeaders headers = new HttpHeaders()
-        headers.setContentType(MediaType.valueOf("application/jsonUTF-8"))
-        HttpEntity<String> strEntity = new HttpEntity<String>(infra, headers)
-
-        List<Long> envList = new ArrayList<>()
-        envList.add(1L)
-        envList.add(2L)
-        envUtil.getConnectedEnvList(_ as EnvListener) >> envList
-        envUtil.getUpdatedEnvList(_ as EnvListener) >> envList
-
-        when:
-        def pages = restTemplate.postForObject("/v1/projects/1/service/list_by_options", strEntity, Page.class)
-
-        then:
-        pages.size() == 1
+        then: '校验返回值'
+        devopsEnvCommandRepository.queryByObject("service", id).getCommandType() == "delete"
     }
 
     def "ListByEnvId"() {
-        given:
+        given: '初始化网络DO类'
         DevopsServiceDO devopsServiceDO = new DevopsServiceDO()
         devopsServiceDO.setEnvId(1L)
         devopsServiceDO.setAppId(1L)
-        devopsServiceDO.setName("svcsvc")
+        devopsServiceDO.setName("svcsvc2")
         devopsServiceDO.setStatus("running")
         devopsServiceDO.setPorts("[{\"port\":7777}]")
         devopsServiceDO.setExternalIp("1.1.1.1")
         devopsServiceDO.setType("ClusterIP")
         devopsServiceDO.setCommandId(1L)
         devopsServiceMapper.insert(devopsServiceDO)
-        when:
+
+        when: '分页查询网络列表'
         def list = restTemplate.getForObject("/v1/projects/1/service?envId=1", List.class)
 
-        then:
+        then: '校验返回值'
         list.size() == 1
     }
 
     def "Query"() {
-        when:
+        when: '查询单个网络'
         def dto = restTemplate.getForObject("/v1/projects/1/service/{id}", DevopsServiceDTO.class, id)
 
-        then:
+        then: '校验返回值'
         dto != null
     }
 
     def "ListByEnv"() {
-        given:
+        given: '设置请求头'
         String infra = "{\"searchParam\":{\"name\":[\"svc\"]}}"
 
         HttpHeaders headers = new HttpHeaders()
@@ -350,17 +315,21 @@ class DevopsServiceControllerSpec extends Specification {
         envUtil.getConnectedEnvList(_ as EnvListener) >> envList
         envUtil.getUpdatedEnvList(_ as EnvListener) >> envList
 
-        when:
+        when: '环境总览网络查询'
         def page = restTemplate.postForObject("/v1/projects/1/service/1/listByEnv", strEntity, Page.class)
 
-        then:
+        then: '校验返回值'
         page.size() == 2
         devopsEnvCommandMapper.deleteByPrimaryKey(1L)
+        devopsEnvCommandMapper.deleteByPrimaryKey(2L)
+        devopsEnvCommandMapper.deleteByPrimaryKey(3L)
+        devopsEnvCommandMapper.deleteByPrimaryKey(4L)
         devopsEnvironmentMapper.deleteByPrimaryKey(1L)
         applicationInstanceMapper.deleteByPrimaryKey(1L)
         devopsEnvFileResourceMapper.deleteByPrimaryKey(1L)
         devopsServiceAppInstanceMapper.deleteByPrimaryKey(1L)
         devopsServiceMapper.deleteByPrimaryKey(id)
+        devopsServiceMapper.deleteByPrimaryKey(2L)
         FileUtil.deleteDirectory(new File("gitops"))
     }
 }
